requirements.txt
starlette>=0.30.0
pydantic>=2
rich>=13
pytest>=8
torch>=2
pandas
tqdm
numpy>=1
setuptools>=68
faker
ollama
google-generativeai
python-Levenshtein
jellyfish
aiohttp
substrate-interface
bittensor
bittensor-cli # Use latest or desired version
wandb
python-dotenv # Add dotenv for .env file handling
requests
unidecode
geonamescache==3.0.0